apm0013173-css-ra-vmanage/requirements.txt
pip

Django==3.2.3
djangorestframework==3.11.1
drf-yasg==1.17.1
gunicorn

requests
coverage
comment-parser

packaging

pandas
xlsxwriter